[fuchsia] Adds --targets arg for build_fuchsia_artifacts.py (#19109)
The Fuchsia artifact builder didn't allow the end user to add more
packages to build for Fuchsia.  The current default is building only
the platform-specific packages; but packages like
`flutter/shell/common:shell_tests` are also defined for Fuchsia but
not compiled by default.

This change makes it less of a hassle to compile such packages.