[Impeller] bump image count to 3 (#52650)
From testing on a Pixel 7 pro, we get constant destruction of buffers at 2, but not at 3, 4, or 5.

I suspect the problem is that when we stop pumping frames, we need to be able to hold all buffers that were in use. Since we support 2 frames in flight while one is encoding, that gives us 3 maximum buffers.