Reverts "macOS: refactor create_macos_framework.py (#54546)" (#54549)
Reverts: flutter/engine#54546
Initiated by: zanderso
Reason for reverting: Failing roll https://github.com/flutter/flutter/pull/153406
Original PR Author: cbracken

Reviewed By: {zanderso}

This change reverts the following previous change:
This is a refactoring with no semantic changes.

This refactors the macOS framework creation code to be more readable, and extracts it to sky_utils.py.

While I was pulling this out, also generalised the code to not hardcode FlutterMacOS.framework in case we one day manage to generate the iOS and macOS frameworks with the same name.

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style