Clean up macOS framework creation scripts (#54658)
This refactors `create_fat_macos_framework`, `_regenerate_symlinks` and
`_set_framework_permissions` to be more generic and not hardcode "FlutterMacOS"
as a framework name. Further, it reuses several utility functions from the iOS
code in `sky_utils` to improve readability and eliminate duplication.

This is refactoring prior to embedding dSYMs in FlutterMacOS.xcframework in a
followup patch.

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style